feat: implemented cooldown, damage, mana cost, projectile count, projection speed and projectile spread upgrades fix: changed AttributeModifierEffect to hold double instead of int to support multipliers refactor: removed vestiges of previous upgrade system refactor: consolidated some registration code refactor: improved application of player attributes perf: ability modifiers are now applied at the beginning of ability processing and removed at the end

+++ src/main/java/com/wanderersoftherift/wotr/modifier/effect/AbstractModifierEffect.java
@@ -19,9 +19,9 @@ import static com.wanderersoftherift.wotr.init.ModModifierEffects.MODIFIER_TYPE_
 public abstract class AbstractModifierEffect {
     public abstract MapCodec<? extends AbstractModifierEffect> getCodec();
 
-    public abstract void enableModifier(float roll, Entity entity,  ModifierSource source);
+    public abstract void enableModifier(double roll, Entity entity,  ModifierSource source);
 
-    public abstract void disableModifier(float roll, Entity entity, ModifierSource source);
+    public abstract void disableModifier(double roll, Entity entity, ModifierSource source);
 
     public abstract void applyModifier();
